+++ lib/eve_online.rb
@@ -53,6 +53,7 @@ require 'eve_online/esi/server_status'
 require 'eve_online/esi/races'
 require 'eve_online/esi/bloodlines'
 require 'eve_online/esi/factions'
+require 'eve_online/esi/ancestries'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'