+++ CHANGELOG.md
@@ -54,6 +54,7 @@
 * Remove `EveOnline::SDE::Models::InvName`
 * Remove `EveOnline::SDE::Models::InvPosition`
 * Add `EveOnline::ESI::Models::Ancestry`
+* Add `EveOnline::ESI::Ancestries`
 
 **v0.14.0**